Changed 'int id' => 'long id' in calculate function signature (t4v4.0 upgrade).

+++ hello_world/objects/hello.c
@@ -11,7 +11,7 @@
 
 using namespace std;
 
-int calculate_hello( obj_loc_t obj_loc, int id )
+int calculate_hello( obj_loc_t obj_loc, long id )
 {
     // Does nothing.
